Move core tests to src
