+++ packages/backend/package.json
@@ -10,7 +10,7 @@
     "format:fix": "prettier --write .",
     "format": "prettier --check .",
     "lint:fix": "yarn lint --fix",
-    "lint": "eslint --ext .ts,.tsx --max-warnings 0 src test",
+    "lint": "eslint --ext .ts,.tsx --max-warnings 0 src",
     "start:dev": "node -r esbuild-register src server",
     "start": "node -r source-map-support/register build/src/index.js server",
     "discover": "node -r esbuild-register src discover",
