+++ packages/backend/src/core/Clock.test.ts
@@ -2,7 +2,7 @@ import { UnixTime } from '@l2beat/types'
 import { install, InstalledClock } from '@sinonjs/fake-timers'
 import { expect } from 'earljs'
 
-import { Clock } from '../../src/core/Clock'
+import { Clock } from './Clock'
 
 describe(Clock.name, () => {
   let time: InstalledClock
