+++ packages/backend/src/core/balances/getBalanceConfigHash.test.ts
@@ -8,190 +8,155 @@ import {
 } from '@l2beat/types'
 import { expect } from 'earljs'
 
-import { getReportConfigHash } from '../../../src/core/reports/getReportConfigHash'
-import { ReportProject } from '../../../src/core/reports/ReportProject'
-import { ProjectEscrow } from '../../../src/model'
+import { ProjectEscrow } from '../../model'
+import { BalanceProject } from './BalanceProject'
+import { getBalanceConfigHash } from './getBalanceConfigHash'
 
-describe(getReportConfigHash.name, () => {
+describe(getBalanceConfigHash.name, () => {
   it('hash changes if project added', () => {
     const projectsBefore = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('dai', 123), fakeToken('eth', 0)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
       ]),
     ]
     const projectsAfter = [
       ...projectsBefore,
-      project('optimism', 'layer2', [
-        fakeEscrow('cc', 2000, [fakeToken('dai', 123)]),
-      ]),
+      project('optimism', [fakeEscrow('cc', 2000, [fakeToken('dai', 123)])]),
     ]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
+    const hashBefore = getBalanceConfigHash(projectsBefore)
+    const hashAfter = getBalanceConfigHash(projectsAfter)
     expect(hashBefore).not.toEqual(hashAfter)
   })
 
   it('hash changes if project is removed', () => {
     const projectsBefore = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('dai', 123), fakeToken('eth', 0)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
       ]),
-      project('optimism', 'layer2', [
-        fakeEscrow('cc', 2000, [fakeToken('dai', 123)]),
-      ]),
+      project('optimism', [fakeEscrow('cc', 2000, [fakeToken('dai', 123)])]),
     ]
     const projectsAfter = [projectsBefore[0]]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
-    expect(hashBefore).not.toEqual(hashAfter)
-  })
-
-  it('hash changes if project type changes', () => {
-    const projectsBefore = [
-      project('unknown', 'layer2', [
-        fakeEscrow('aa', 1000, [fakeToken('dai', 123)]),
-      ]),
-    ]
-    const projectsAfter = [
-      project('unknown', 'bridge', [
-        fakeEscrow('aa', 1000, [fakeToken('dai', 123)]),
-      ]),
-    ]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
+    const hashBefore = getBalanceConfigHash(projectsBefore)
+    const hashAfter = getBalanceConfigHash(projectsAfter)
     expect(hashBefore).not.toEqual(hashAfter)
   })
 
   it('hash changes if token is added', () => {
     const projectsBefore = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('dai', 123), fakeToken('eth', 0)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
       ]),
     ]
     const projectsAfter = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('dai', 123), fakeToken('eth', 0)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123), fakeToken('usdc', 456)]),
       ]),
     ]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
+    const hashBefore = getBalanceConfigHash(projectsBefore)
+    const hashAfter = getBalanceConfigHash(projectsAfter)
     expect(hashBefore).not.toEqual(hashAfter)
   })
 
   it('hash changes if token is removed', () => {
     const projectsBefore = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('dai', 123), fakeToken('eth', 0)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
       ]),
     ]
     const projectsAfter = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('eth', 0)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
       ]),
     ]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
+    const hashBefore = getBalanceConfigHash(projectsBefore)
+    const hashAfter = getBalanceConfigHash(projectsAfter)
     expect(hashBefore).not.toEqual(hashAfter)
   })
 
   it('hash changes if escrow sinceBlock changes', () => {
     const projectsBefore = [
-      project('arbitrum', 'layer2', [
-        fakeEscrow('aa', 1000, [fakeToken('dai', 123)]),
-      ]),
+      project('arbitrum', [fakeEscrow('aa', 1000, [fakeToken('dai', 123)])]),
     ]
     const projectsAfter = [
-      project('arbitrum', 'layer2', [
-        fakeEscrow('aa', 2000, [fakeToken('dai', 123)]),
-      ]),
+      project('arbitrum', [fakeEscrow('aa', 2000, [fakeToken('dai', 123)])]),
     ]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
+    const hashBefore = getBalanceConfigHash(projectsBefore)
+    const hashAfter = getBalanceConfigHash(projectsAfter)
     expect(hashBefore).not.toEqual(hashAfter)
   })
 
   it('hash changes if token sinceBlock changes', () => {
     const projectsBefore = [
-      project('arbitrum', 'layer2', [
-        fakeEscrow('aa', 1000, [fakeToken('dai', 123)]),
-      ]),
+      project('arbitrum', [fakeEscrow('aa', 1000, [fakeToken('dai', 123)])]),
     ]
     const projectsAfter = [
-      project('arbitrum', 'layer2', [
-        fakeEscrow('aa', 1000, [fakeToken('dai', 456)]),
-      ]),
+      project('arbitrum', [fakeEscrow('aa', 1000, [fakeToken('dai', 456)])]),
     ]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
+    const hashBefore = getBalanceConfigHash(projectsBefore)
+    const hashAfter = getBalanceConfigHash(projectsAfter)
     expect(hashBefore).not.toEqual(hashAfter)
   })
 
   it('hash stays the same if the project order changes', () => {
     const projectsBefore = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('dai', 123), fakeToken('eth', 0)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
       ]),
-      project('optimism', 'layer2', [
-        fakeEscrow('cc', 2000, [fakeToken('dai', 123)]),
-      ]),
+      project('optimism', [fakeEscrow('cc', 2000, [fakeToken('dai', 123)])]),
     ]
     const projectsAfter = [projectsBefore[1], projectsBefore[0]]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
+    const hashBefore = getBalanceConfigHash(projectsBefore)
+    const hashAfter = getBalanceConfigHash(projectsAfter)
     expect(hashBefore).toEqual(hashAfter)
   })
 
   it('hash stays the same if the escrow order changes', () => {
     const projectsBefore = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('dai', 123), fakeToken('eth', 0)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
       ]),
     ]
     const projectsAfter = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
         fakeEscrow('aa', 1000, [fakeToken('dai', 123), fakeToken('eth', 0)]),
       ]),
     ]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
+    const hashBefore = getBalanceConfigHash(projectsBefore)
+    const hashAfter = getBalanceConfigHash(projectsAfter)
     expect(hashBefore).toEqual(hashAfter)
   })
 
   it('hash stays the same if the token order changes', () => {
     const projectsBefore = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('dai', 123), fakeToken('eth', 0)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
       ]),
     ]
     const projectsAfter = [
-      project('arbitrum', 'layer2', [
+      project('arbitrum', [
         fakeEscrow('aa', 1000, [fakeToken('eth', 0), fakeToken('dai', 123)]),
         fakeEscrow('bb', 2000, [fakeToken('dai', 123)]),
       ]),
     ]
-    const hashBefore = getReportConfigHash(projectsBefore)
-    const hashAfter = getReportConfigHash(projectsAfter)
+    const hashBefore = getBalanceConfigHash(projectsBefore)
+    const hashAfter = getBalanceConfigHash(projectsAfter)
     expect(hashBefore).toEqual(hashAfter)
   })
 })
 
-function project(
-  id: string,
-  type: ReportProject['type'],
-  escrows: ProjectEscrow[],
-): ReportProject {
+function project(id: string, escrows: ProjectEscrow[]): BalanceProject {
   return {
     projectId: ProjectId(id),
-    type,
     escrows,
   }
 }
